requirements.txt
pycoingecko
python-telegram-bot
persiantools
DateTime
pytz
